+++ lib/src/utils/data_cache.dart
@@ -25,7 +25,6 @@ class DataCache {
       var hasData = PrefService.getString(item.key) != null;
       if (hasData) continue;
       var content = await rootBundle.loadString(item.path);
-      print(content);
       PrefService.setString(item.key, content);
     }
   }